Move `+compiler-libs` to its own directive.
This is ok because it happens unconditionally: it doesn't hurt on OCaml
< 4 and does not trigger any printing on the terminal.

It needs to be in a separate phrase because the search path is only
taken into account after the phrase is finished, thus attempting to load
`+compiler-libs` using

let () = Topdirs.dir_directory "+compiler-libs" in
(* something using modules in compiler-libs *)

does not work in the Toplevel. Unfortunately the toplevel is printing
return values so the amount of phrases has to be minimized to avoid
additional prints of

- : unit = ()

for every expression that is evaluated.